% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elevations.R
\name{elevations}
\alias{elevations}
\title{Create Visualisations of Building Elevations}
\usage{
elevations(nbuildings, min_storey, max_storey)
}
\arguments{
\item{nbuildings}{Number of buildings}

\item{min_storey}{Minimum permitted storey per building}

\item{max_storey}{Maximum permitted storey per building}
}
\value{
Small multiple charts depicting building elevations and a print-out of the number of potential combinations in the console
}
\description{
Function for charting building elevations and the number of potential combinations using permutations.
}
\examples{
elevations(nbuildings = 6, min_storey = 4, max_storey = 5)
}
