% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elevations.R
\name{elevations}
\alias{elevations}
\title{Create Visualisations of Building Elevations}
\usage{
elevations(n_buildings, min_storey, max_storey)
}
\arguments{
\item{n_buildings}{Number of buildings}

\item{min_storey}{Minimum permitted storey}

\item{max_storey}{Maximum permitted storey}
}
\value{
Small multiple charts depicting building elevations and the number of potential combinations
}
\description{
Function for charting building elevations and the number of potential combinations using permutations.
}
\examples{
elevations(n_buildings = 7, min_storey = 4, max_storey = 5)
}
