% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extension_elevations.R
\name{extension_elevations}
\alias{extension_elevations}
\title{Create Visualisations of Building Elevations for Rooftop Extensions}
\usage{
extension_elevations(
  storeys = NULL,
  df = NULL,
  max_permitted_storey = NULL,
  max_permitted_extstorey = NULL
)
}
\arguments{
\item{storeys}{Vector containing the number of storeys per building}

\item{df}{Data frame of existing elevations. Each column should represent a building and the data frame should contain one row with the number of storeys per building}

\item{max_permitted_storey}{Maximum permitted storey per building}

\item{max_permitted_extstorey}{Maximum permitted storey per extension}
}
\value{
Multiple bar charts depicting potential rooftop extension elevations ordered by density as well as a print-out of the number of potential combinations in the console
}
\description{
Function for charting building elevation small multiples for rooftop extensions and the number of potential combinations using permutations.
}
\examples{
# Visualise potentials based on the maximum number of storeys per extension
storeys_i = c(3, 3, 4)
extension_elevations(storeys = storeys_i, max_permitted_extstorey = 2)

# Chart potential extensions using the maximum number of storeys per building
storeys_ii = c(3, 2, 2, 3, 4)
extension_elevations(storeys = storeys_ii, max_permitted_storey = 4)
}
