% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extension_elevations.R
\name{extension_elevations}
\alias{extension_elevations}
\title{Create Permutations of Potential Rooftop Extension Elevations}
\usage{
extension_elevations(
  storeys = NULL,
  df = NULL,
  max_permitted_storey = NULL,
  max_permitted_extstorey = NULL,
  output = "plot"
)
}
\arguments{
\item{storeys}{Vector containing the number of storeys per building}

\item{df}{Data frame of existing elevations. Each column should represent a building and the data frame should contain one row with the number of storeys per building}

\item{max_permitted_storey}{Maximum permitted storey per building}

\item{max_permitted_extstorey}{Maximum permitted storey per extension}

\item{output}{Output of function. Input 'plot' for charts. Input 'data' for a data frame containing the number of storeys per building. Defaults to 'plot'}
}
\value{
Multiple charts depicting potential rooftop extension elevations ordered by density as well as a print-out of the number of potential combinations in the console or a data frame containing the number of storeys per building
}
\description{
Function for generating visualisations of permutations of rooftop extension elevations and the number of potential combinations or a data frame containing storeys per building.
}
\examples{
storeys_i = c(3, 3, 4) # Establish the number of storeys per existing building

# Visualise potentials based on the maximum number of storeys per extension
extension_elevations(storeys = storeys_i, max_permitted_extstorey = 2, output = 'plot')

# Generate a dataset containing permutations of storeys per building
extension_elevations(storeys = storeys_i, max_permitted_extstorey = 1, output = 'data')

storeys_ii = c(3, 2, 2, 3, 4) # Establish the number of storeys per existing building

# Chart potential extensions using the maximum number of storeys per building
extension_elevations(storeys = storeys_ii, max_permitted_storey = 4, output = 'plot')
}
