% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extension_elevations.R
\name{extension_elevations}
\alias{extension_elevations}
\title{Building Elevations for Potential Rooftop Extensions}
\usage{
extension_elevations(df, max_permitted_storey)
}
\arguments{
\item{df}{Data frame containing building elevations including the number of storeys}

\item{max_permitted_storey}{Maximum permitted storey}
}
\value{
Small multiple charts depicting rooftop extension elevations according to the maximum storeys permitted and the number of possible combinations in the console
}
\description{
Function for creating building elevation small multiples for potential rooftop extensions using permutations.
}
\examples{
extension_elevations(df = df, max_permitted_storey = 5)
}
